refactor(launcher): Move property logging functions
Fit better in already existing property module
